Checkpoint

+++ src/client/explorer.rs
@@ -28,7 +28,7 @@ impl From<Money> for ExplorerMoney {
 
 #[derive(Deserialize, Serialize, Debug, Clone)]
 pub struct ExplorerMpnAccount {
-    pub nonce: u64,
+    pub nonce: u32,
     pub address: String,
     pub tokens: HashMap<u64, ExplorerMoney>,
 }
@@ -481,7 +481,7 @@ pub struct ExplorerMpnWithdraw {
     pub zk_address: String,
     pub zk_token_index: u64,
     pub zk_fee_token_index: u64,
-    pub zk_nonce: u64,
+    pub zk_nonce: u32,
     pub zk_sig: String,
     pub payment: ExplorerContractWithdraw,
 }
@@ -501,7 +501,7 @@ impl From<&MpnWithdraw> for ExplorerMpnWithdraw {
 
 #[derive(Deserialize, Serialize, Debug, Clone)]
 pub struct ExplorerMpnTransaction {
-    pub nonce: u64,
+    pub nonce: u32,
     pub src_pub_key: String,
     pub dst_pub_key: String,
